[gtk] Fix three CodeQL "multiplication result converted to larger type" warnings.
Use static_cast<size_t>() and similar to fix it.

+++ src/gtk/PIMGTYPE.cpp
@@ -105,7 +105,8 @@ PIMGTYPE PIMGTYPE_scale(PIMGTYPE pImgType, int width, int height, bool bilinear)
 	// NOTE: GdkMemoryTexture only does a g_bytes_ref() if the stride matches
 	// what it expects, so we need to do a deep copy here.
 	const int dest_stride = cairo_image_surface_get_stride(dest_surface);
-	GBytes *const pBytes = g_bytes_new(cairo_image_surface_get_data(dest_surface), height * dest_stride);
+	GBytes *const pBytes = g_bytes_new(cairo_image_surface_get_data(dest_surface),
+		static_cast<gsize>(height) * static_cast<gsize>(dest_stride));
 	// FIXME: GDK_MEMORY_DEFAULT (GDK_MEMORY_B8G8R8A8_PREMULTIPLIED) causes a heap overflow here...
 	GdkTexture *const texture = gdk_memory_texture_new(width, height, GDK_MEMORY_B8G8R8A8, pBytes, dest_stride);
 	g_bytes_unref(pBytes);
